Fix can't remove unsaved teacher

+++ src/js/components/teach-setting.vue
@@ -64,6 +64,11 @@ module.exports = {
     },
     removeTeacher: function(id) {
       var vm = this;
+      if (!id && vm.selectedIndex > -1) {
+        // remove unsaved teacher after creating
+        vm.data.splice(vm.selectedIndex, 1);
+        return;
+      }
       var request = teacher_service.remove(id);
       request.done(function(data, textStatus, jqXHR) {
         if (data.n === 1 && data.ok === 1) {
